feat: unified git diff get directories
